Lab2_DigitDetection/DETR/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
Pillow>=10.0.0
PyYAML>=6.0.1
pycocotools>=2.0.7
scipy>=1.10.0
tqdm>=4.66.0
gdown>=4.7.1
tensorboard>=2.16.2
